Removed cycled dependencies, added enum
